+++ contrib/platform/src/com/sun/jna/platform/win32/COM/tlb/imp/TlbPropertyPut.java
@@ -53,8 +53,8 @@ public class TlbPropertyPut extends TlbAbstractMethod implements Variant {
         this.methodName = "set" + getMethodName();
         String methodparams = "";
         String methodvariables = ", ";
-        int vtableId = funcDesc.oVft;
-        int paramCount = funcDesc.cParams;
+        short vtableId = funcDesc.oVft.shortValue();
+        short paramCount = funcDesc.cParams.shortValue();
         String varType;
         String[] names = typeInfoUtil.getNames(funcDesc.memid, paramCount + 1);
